+++ src/components/common/MainTitle/index.tsx
@@ -1,15 +1,5 @@
 import { PropsWithChildren } from "react";
-import styled from "@emotion/styled";
-
-const S = {
-  Title: styled.h1`
-    font-size: 24px;
-    font-weight: 700;
-    line-height: 36px;
-
-    margin: 10px 0;
-  `,
-};
+import S from "./styledComponent";
 
 const MainTitle = ({ children }: PropsWithChildren) => {
   return <S.Title>{children}</S.Title>;
