+++ src/components/common/MainTitle/styledComponent.ts
@@ -1,0 +1,13 @@
+import styled from "@emotion/styled";
+
+const S = {
+  Title: styled.h1`
+    font-size: 24px;
+    font-weight: 700;
+    line-height: 36px;
+
+    margin: 10px 0;
+  `,
+};
+
+export default S;
