+++ src/components/product/ToggleItemButton/index.tsx
@@ -1,38 +1,14 @@
-import styled from "@emotion/styled";
 import { useEffect, useState } from "react";
 
-import COLOR_PALETTE from "../../style/colorPalette";
+import useCustomContext from "../../../hooks/useCustomContext";
 
-import useCustomContext from "../../hooks/useCustomContext";
+import { ToggleCartItemContext } from "../../provider/ToggleCartItemProvider";
 
-import { ToggleCartItemContext } from "../ToggleCartItemProvider";
+import AddToCart from "../../icons/AddToCart";
+import DeleteFromCart from "../../icons/DeleteFromCart";
+import LoadingDots from "../../common/LoadingDots";
 
-import AddToCart from "../icons/AddToCart";
-import DeleteFromCart from "../icons/DeleteFromCart";
-import LoadingDots from "../LoadingDots";
-
-const S = {
-  ToggleItemButton: styled.button<{ isSelected: boolean }>`
-    width: 60px;
-    height: 32px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    gap: 4px;
-
-    background-color: ${({ isSelected }) => (isSelected ? COLOR_PALETTE.lightGrey : COLOR_PALETTE.black)};
-    color: ${({ isSelected }) => (isSelected ? COLOR_PALETTE.black : COLOR_PALETTE.white)};
-    padding: 8px;
-    margin: 4px;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    align-self: flex-end;
-    &:hover {
-      opacity: 0.8;
-    }
-  `,
-};
+import S from "./StyledComponent";
 
 interface ToggleItemButtonProps {
   id: number;
