+++ src/components/utils/InfiniteScroll/styledComponent.ts
@@ -1,0 +1,10 @@
+import styled from "@emotion/styled";
+
+const S = {
+  FallbackContainer: styled.div`
+    width: calc(183px * 2 + 16px);
+    margin: 16px 0;
+  `,
+};
+
+export default S;
