+++ src/pages/Mall/index.tsx
@@ -1,44 +1,24 @@
 import { useEffect } from "react";
 import { Global } from "@emotion/react";
-import styled from "@emotion/styled";
 
-import { baseStyle } from "../style/baseStyle";
+import { baseStyle } from "../../style/baseStyle";
 
-import useProducts from "../hooks/useProducts";
-import useCustomContext from "../hooks/useCustomContext";
+import useProducts from "../../hooks/useProducts";
+import useCustomContext from "../../hooks/useCustomContext";
 
-import { ToggleCartItemContext } from "../components/ToggleCartItemProvider";
-import { ToastContext } from "../components/Toasts/ToastProvider";
+import { ToggleCartItemContext } from "../../components/provider/ToggleCartItemProvider";
+import { ToastContext } from "../../components/provider/ToastProvider";
 
-import Dropdown from "../components/Dropdown";
-import Header from "../components/Header";
-import InfiniteScrollComponent from "../components/InfiniteProductsScrollComponent";
-import MainTitle from "../components/MainTitle";
-import ProductCard from "../components/product/ProductCard";
+import Dropdown from "../../components/common/Dropdown";
+import Header from "../../components/layout/Header";
+import InfiniteScroll from "../../components/utils/InfiniteScroll";
+import MainTitle from "../../components/common/MainTitle";
+import ProductCard from "../../components/product/ProductCard";
 
-import { PRODUCT_CATEGORY, PRODUCT_SORT } from "../constants/mall";
-import { ERROR_MESSAGE } from "../constants/message";
+import { PRODUCT_CATEGORY, PRODUCT_SORT } from "../../constants/mall";
+import { ERROR_MESSAGE } from "../../constants/message";
 
-const S = {
-  MainMall: styled.div`
-    padding: 36px 24px;
-
-    display: flex;
-    flex-direction: column;
-    gap: 24px;
-  `,
-
-  Toolbar: styled.div`
-    display: flex;
-    justify-content: space-between;
-  `,
-
-  ProductList: styled.div`
-    display: grid;
-    grid-template-columns: repeat(2, 183px);
-    gap: 16px;
-  `,
-};
+import S from "./styledComponent";
 
 const Mall = () => {
   const {
@@ -77,12 +57,12 @@ const Mall = () => {
           <Dropdown options={PRODUCT_SORT} handleChange={handleSortChange} />
         </S.Toolbar>
         <S.ProductList>
-          <InfiniteScrollComponent isLoading={isProductLoading} handleScroll={fetchNextPage}>
+          <InfiniteScroll isLoading={isProductLoading} handleScroll={fetchNextPage}>
             {products.map((product, index) => (
               <ProductCard key={`${index}${product.id}`} product={product} />
             ))}
             {productError! && <div>{ERROR_MESSAGE.getProducts}</div>}
-          </InfiniteScrollComponent>
+          </InfiniteScroll>
         </S.ProductList>
       </S.MainMall>
     </>
